pyproject.toml
[tool.poetry]
name = "talk-bot"
version = "0.1.0"
description = ""
authors = ["John Victor <johnvictorfs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
peewee = "^3.8"
psycopg2-binary = "^2.7"
"discord.py" = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
